refactor: app_lock module
