+++ packages/cryptoplease/lib/features/app_lock/src/presentation/app_lock_disable_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../l10n/l10n.dart';
-import '../../../ui/decorated_window/decorated_window.dart';
+import '../../../../l10n/l10n.dart';
+import '../../../../ui/decorated_window/decorated_window.dart';
 import '../bl/app_lock_bloc.dart';
 import 'app_lock_setup_flow_screen.dart';
 import 'components/pin_input_display_widget.dart';
