+++ packages/cryptoplease/lib/features/app_lock/src/presentation/app_lock_settings.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../l10n/l10n.dart';
-import '../../../ui/app_bar.dart';
-import '../../../ui/switch.dart';
-import '../../../ui/theme.dart';
+import '../../../../l10n/l10n.dart';
+import '../../../../ui/app_bar.dart';
+import '../../../../ui/switch.dart';
+import '../../../../ui/theme.dart';
 import '../bl/app_lock_bloc.dart';
 import 'app_lock_setup_flow_screen.dart';
