+++ packages/cryptoplease/lib/features/app_lock/src/presentation/components/pin_display.dart
