+++ packages/cryptoplease/lib/features/app_lock/src/presentation/components/pin_keypad.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../../ui/amount_keypad/keypad_key.dart';
+import '../../../../../ui/amount_keypad/keypad_key.dart';
 
 class PinKeypad extends StatelessWidget {
   const PinKeypad({
